+++ src/View/Personalizations/Support/Colors/ErrorsColors.php
@@ -3,8 +3,8 @@
 namespace TallStackUi\View\Personalizations\Support\Colors;
 
 use TallStackUi\Facades\TallStackUi;
-use TallStackUi\Support\Color;
 use TallStackUi\View\Components\Errors;
+use TallStackUi\View\Personalizations\Support\Color;
 
 class ErrorsColors
 {
