+++ src/View/Personalizations/Support/Colors/ToggleColors.php
@@ -3,8 +3,8 @@
 namespace TallStackUi\View\Personalizations\Support\Colors;
 
 use TallStackUi\Facades\TallStackUi;
-use TallStackUi\Support\Color;
 use TallStackUi\View\Components\Form\Toggle;
+use TallStackUi\View\Personalizations\Support\Color;
 
 class ToggleColors
 {
